+++ app/src/main/java/com/boostcamp/dreampicker/data/repository/FeedRepository.java
@@ -21,15 +21,4 @@ public interface FeedRepository {
     // [업로드] 사용자가 작성한 피드를 서버에 업로드한다.
     @NonNull
     Completable uploadFeed(@NonNull final FeedUploadRequest feed);
-  
-    /**
-     * 유저가 업로드 한 투표 목록 페이징
-     *
-     * @param userId 유저 ID
-     * @param startAfter 이 값의 다음 데이터를 요청
-     * @param pageSize 페이지의 size
-     * @return 피드 리스트 stream
-     */
-    @NonNull
-    Single<List<MyFeed>> getFeedListByUserId(@NonNull final String userId, final Date startAfter, final int pageSize);
 }
